backend/requirements.txt
fastapi
uvicorn
python-multipart
pillow
torch
accelerate
transformers
huggingface_hub
PyMuPDF
python-dotenv
docling-core 
